Split UTs for freestyle and pipeline jobs
